refactor: Make `tls13server` a library function.
Note: This is used in the next commit to implement the self-test.
